One option-table matcher: merge OptionTable into prefix as the single API
The two deep reviews (#929 and this branch) independently extracted the
same Tcl_GetIndexFromObjStruct port — #929 as free functions over
AsRef<[u8]> tables (crate::prefix), this branch as a const-constructible
typed table (OptionTable). Rather than layering one over the other, the
two merge: prefix::OptionTable is now generic over AsRef<[u8]> entries
(a static &str table, the VM's runtime String table, and the runtimes'
byte tables all resolve through the same value) and owns the resolve/
index_of/index_of_str surface; the raw lookup/lookup_error free
functions go private. The composing escape hatches stay public for
sites that build their own sentence: scan (the noun-free rule),
choice_list/choice_list_bytes, and bad_key_message (byte nouns — the
WASM runtime's tcl::prefix match -message). Every former lookup()
caller (VM tcl::prefix, runtime string ensemble/OO tables) ports to
the merged API; option_table.rs is deleted and its tests fold into
prefix.rs. The shared-utility contract doc describes the single
module.
