+++ rust/tcl-cmd-core/src/lib.rs
@@ -49,7 +49,6 @@ pub mod lseq;
 pub mod lsort;
 pub mod mathop;
 pub mod namespace;
-pub mod option_table;
 pub mod path;
 pub mod platform;
 pub mod prefix;
